robustness_benchmark/requirements.txt
av>=10.0.0
numpy>=1.19.5
pandas>=1.1.5
Pillow>=10.0.0
scikit_image>=0.17.2
scikit_learn>=1.3.0
scipy>=1.4.1
setuptools>=59.5.0
torchvision>=0.11.3
tqdm>=4.57.0
torch>=1.10.2
opencv_python>=4.2.0.32
